feat: Add error handler

+++ api/review-api.js
@@ -10,11 +10,27 @@ router.get("/reviews", async (req, res) => {
 
 router.post("/reviews", async (req, res, next) => {
     try {
-        const checkReview = await Review.findOne({ shop: req.body.shop });
+        if (req.body.shopName === undefined) {
+            throw new InputError("'shopName' parameter is empty.");
+        }
+        if (req.body.author === undefined) {
+            throw new InputError("'author parameter is empty.");
+        }
+
+        if (req.body.date === undefined) {
+            review.date = getToday();
+        } else {
+            review.date = req.body.date;
+        }
+
+        const checkReview = await Review.findOne({ shopName: req.body.shopName });
+
+        if (checkReview) {
+            throw new DuplicateError();
+        }
         const review = new Review();
         review.shopName = req.body.shopName;
         review.author = req.body.author;
-        review.date = req.body.date;
         review.description = req.body.description;
         review.keyword = req.body.keyword;
 
@@ -27,6 +43,9 @@ router.post("/reviews", async (req, res, next) => {
 
 router.get("/reviews/:id", async (req, res, next) => {
     try {
+        if (isNaN(parseInt(req.params.id))) {
+            throw new InputError("Invalid review id.");
+        }
         const targetReview = await Review.findOne({ id: req.params.id });
         if (!targetReview) {
             throw new NotFoundError();
@@ -39,6 +58,21 @@ router.get("/reviews/:id", async (req, res, next) => {
 
 router.put("/reviews/:id", async (req, res, next) => {
     try {
+        if (isNaN(parseInt(req.params.id))) {
+            throw new InputError("Invalid review id.");
+        }
+        if (req.body.shopName === undefined) {
+            throw new InputError("'shopName' parameter is empty.");
+        }
+        if (req.body.author === undefined) {
+            throw new InputError("'author parameter is empty.");
+        }
+        if (req.body.date === undefined) {
+            review.date = getToday();
+        } else {
+            review.date = req.body.date;
+        }
+
         const targetReview = await Review.findOne({ id: req.params.id });
         if (!targetReview) {
             throw new NotFoundError();
@@ -58,7 +92,13 @@ router.put("/reviews/:id", async (req, res, next) => {
 
 router.delete("/reviews/:id", async (req, res, next) => {
     try {
+        if (isNaN(parseInt(req.params.id))) {
+            throw new InputError("Invalid review id.");
+        }
         const targetReview = await Review.findOne({ id: req.params.id });
+        if (!targetReview) {
+            throw new NotFoundError();
+        }
         const deleteReview = await Review.deleteOne({ id: req.params.id });
         if (deleteReview.deletedCount === 1) {
             res.status(200).send({ id: targetReview.id, shopName: targetReview.shopName, author: targetReview.author });
@@ -74,7 +114,16 @@ router.use(function (err, req, res, next) {
     if (res.headersSent) {
         return next(err);
     }
-    res.status(404).json({ message: err.message });
+    res.status(err.status).json({ message: err.message });
 });
+
+function getToday() {
+    const today = new Date();
+    const year = today.getFullYear();
+    const month = ("0" + (1 + today.getMonth())).slice(-2);
+    const day = ("0" + today.getDate()).slice(-2);
+
+    return year + "-" + month + "-" + day;
+}
 
 module.exports = router;
